+++ README.md
@@ -58,7 +58,7 @@ Finally, you feed it the worker script and a list of `Task` objects to execute o
 The `worker` will most likely be the very core of your requestor app. You need to define
 this function in your agent code and then you pass it to the Executor.
 
-It receives a `WorkContext` (`yapapi.runner.ctx.WorkContext`) object that serves 
+It receives a `WorkContext` (`yapapi.WorkContext`) object that serves 
 as an interface between your script and the execution unit within the provider. 
 Using the work context, you define the steps that the provider needs to execute in order
 to complete the job you're giving them - e.g. transferring files to and from the provider
@@ -75,7 +75,7 @@ processing of the file using a set of parameters specified in the `Task` data.
 
 #### Task
 
-The `Task` (`yapapi.runner.task.Task`) object that describes a fragment of your task,
+The `Task` (`yapapi.Task`) object that describes a fragment of your task,
 that is a single piece of your application's job that will be executed in a single run
 of the execution script on a provider's machine.
