+++ tests/executor/test_ctx.py
@@ -1,4 +1,4 @@
-from yapapi.runner.ctx import CommandContainer
+from yapapi.executor.ctx import CommandContainer
 import json
