+++ yapapi/__init__.py
@@ -4,7 +4,7 @@ import toml
 from pathlib import Path
 from pkg_resources import get_distribution
 
-from .runner import Executor, Task, WorkContext
+from .executor import Executor, Task, WorkContext
 
 
 def get_version() -> str:
@@ -22,4 +22,4 @@ def get_version() -> str:
 
 
 __version__: str = get_version()
-__all__ = ["Executor", "props", "rest", "runner", "storage", "Task", "WorkContext"]
+__all__ = ["Executor", "props", "rest", "executor", "storage", "Task", "WorkContext"]
