+++ yapapi/executor/strategy.py
