+++ yapapi/log.py
@@ -4,7 +4,7 @@ Functions in this module fall into two categories:
 
 * Functions that convert computation events generated by the `Executor.submit`
 method to calls to the standard Python `logging` module, using the loggers
-in the "yapapi" namespace (e.g. `logging.getLogger("yapapi.runner")`).
+in the "yapapi" namespace (e.g. `logging.getLogger("yapapi.executor")`).
 These functions should be passed as `event_emitter` arguments to `Executor()`.
 
 * Functions that perform configuration of the `logging` module itself.
@@ -18,7 +18,7 @@ Functions for handling events
 -----------------------------
 
 Several functions from this module can be passed as `event_emitter` callback to
-`yapapi.runner.Executor()`.
+`yapapi.executor.Executor()`.
 
 For detailed, human-readable output use the `log_event` function:
 ```python
@@ -49,10 +49,10 @@ import logging
 import time
 from typing import Any, Callable, Dict, Iterator, List, Optional, Set
 
-import yapapi.runner.events as events
+import yapapi.executor.events as events
 
 
-logger = logging.getLogger("yapapi.runner")
+logger = logging.getLogger("yapapi.executor")
 
 
 def enable_default_logger(
